+++ packages/react-sdk/src/components/components.ts
@@ -1,19 +1,41 @@
-export { Body } from "./body";
-export { Box } from "./box";
-export { TextBlock } from "./text-block";
-export { Heading } from "./heading";
-export { Paragraph } from "./paragraph";
-export { Link } from "./link";
-export { RichTextLink } from "./rich-text-link";
-export { Span } from "./span";
-export { Bold } from "./bold";
-export { Italic } from "./italic";
-export { Superscript } from "./superscript";
-export { Subscript } from "./subscript";
-export { Button } from "./button";
-export { Input } from "./input";
-export { Form } from "./form";
-export { Image } from "./image";
-export { Blockquote } from "./blockquote";
-export { List } from "./list";
-export { ListItem } from "./list-item";
+import { Body } from "./body";
+import { Box } from "./box";
+import { TextBlock } from "./text-block";
+import { Heading } from "./heading";
+import { Paragraph } from "./paragraph";
+import { Link } from "./link";
+import { RichTextLink } from "./rich-text-link";
+import { Span } from "./span";
+import { Bold } from "./bold";
+import { Italic } from "./italic";
+import { Superscript } from "./superscript";
+import { Subscript } from "./subscript";
+import { Button } from "./button";
+import { Input } from "./input";
+import { Form } from "./form";
+import { Image } from "./image";
+import { Blockquote } from "./blockquote";
+import { List } from "./list";
+import { ListItem } from "./list-item";
+
+export {
+  Body,
+  Box,
+  TextBlock,
+  Heading,
+  Paragraph,
+  Link,
+  RichTextLink,
+  Span,
+  Bold,
+  Italic,
+  Superscript,
+  Subscript,
+  Button,
+  Input,
+  Form,
+  Image,
+  Blockquote,
+  List,
+  ListItem,
+};
