+++ packages/sdk-size-test/app/page-components/page-main.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/no-internal-modules
+import { Box, Button, Body, createGetComponent } from "@webstudio-is/react-sdk";
+
+export const getComponent = createGetComponent({ Box, Button, Body });
